Message path autocomplete uses an unconstrained popup (#7353)
The current autocomplete component for message path sets the width of
the suggestion list to the width of the input. When used on the settings
sidebar with long paths this will truncate the topic names and
suggestions. Instead of limiting the width of the input, we can allow it
to use more space on the page.

This change uses a custom popper component for the MuiAutocomplete to
set a `minWidth` property rather than `width` as the built-in behavior
does. I've also updated the `renderOption` logic to create a tuple
rather than rendering the `MenuItem` component. Since we use a
virtualized list, we let the virtualization logic determine when to
render the component since it only renders some of the elements in the
list avoid building a MenuItem instance for all items since many won't
be visible.

Co-authored-by: Andrew Carr <andrew@foxglove.dev>
Co-authored-by: Jacob Bandes-Storch <jacob@foxglove.dev>