(fix) handle empty min/maxYValue strings (#7417)
**User-Facing Changes**
<!-- will be used as a changelog entry -->
- FIX: Empty plot configuration string values for minYValue and
maxYValue led to an initially rendered Y scale of 0-0

**Description**
The minYValue and maxYValue type currently supports number _or_ string,
but empty strings were cast to 0 when creating the Y-bounds on the Plot
for demo data. This led to y bounds of min: 0 and max: 0 (rather than
defaulting to auto scale). I removed the empty string values from the
test data and added a Storybook case to cover backwards compatibility.

<!-- link relevant GitHub issues -->
<!-- add `docs` label if this PR requires documentation updates -->
<!-- add relevant metric tracking for experimental / new features -->