(fix) prevent play button flicker (#7423)
**User-Facing Changes**
<!-- will be used as a changelog entry -->
- FIX: Prevent play button flicker on rapidly repeating playback

**Description**
The controls for repeating playback have been moved from the
RepeatAdaptor to the iterablePlayer. This allows us to more directly
determine the order of operations for changing the isPlaying/idle states
as they relate to repeating the data. This change means the 'isPlaying'
state remains consistent while the data loops, which prevents the odd
play button flashing behavior.

<!-- link relevant GitHub issues -->
<!-- add `docs` label if this PR requires documentation updates -->
<!-- add relevant metric tracking for experimental / new features -->